refactor: move search out of Header into OperationsScreen
Header is now a minimal rightContent pass-through with no search logic.
SearchBar, FilterChipStrip, SearchOverlay, and all open/close/collapse
state now live entirely in OperationsScreen. Removes leftover debug
console.log statements. Tests updated to match the new architecture.

🧀
