+++ __tests__/components/Header-rightContent.test.js
@@ -8,34 +8,14 @@ jest.mock('../../app/contexts/ThemeColorsContext', () => ({
   useThemeColors: () => ({
     colors: {
       background: '#fff',
-      border: '#ddd',
-      text: '#000',
-      mutedText: '#999',
-      primary: '#007AFF',
     },
   }),
 }));
 
-jest.mock('../../app/contexts/LocalizationContext', () => ({
-  useLocalization: () => ({ t: (key) => key }),
-}));
-
-jest.mock('../../app/contexts/UpdateDownloadContext', () => ({
-  useUpdateDownload: () => ({ isDownloading: false, downloadProgress: 0 }),
-}));
-
-jest.mock('../../app/contexts/SearchContext', () => ({
-  useSearch: () => ({ openSearch: jest.fn(), registerSearchHandler: jest.fn() }),
-}));
-
-jest.mock('../../app/services/db', () => ({
-  getDatabaseVersion: jest.fn().mockResolvedValue(1),
-}));
-
 describe('Header - rightContent prop', () => {
-  it('renders without crashing when rightContent not provided', () => {
+  it('returns null when rightContent not provided', () => {
     const { toJSON } = render(<Header />);
-    expect(toJSON()).toBeTruthy();
+    expect(toJSON()).toBeNull();
   });
 
   it('renders rightContent when provided', () => {
