+++ app/screens/OperationsScreen.js
@@ -1,5 +1,5 @@
 import React, { useState, useMemo, useCallback, useEffect, useRef } from 'react';
-import { View, StyleSheet, FlatList, TouchableOpacity, TextInput, Pressable, Modal, Keyboard, InteractionManager } from 'react-native';
+import { View, StyleSheet, FlatList, TouchableOpacity, TextInput, Pressable, Modal, Keyboard, InteractionManager, BackHandler } from 'react-native';
 import Animated, { useSharedValue, useAnimatedStyle, withTiming, Easing } from 'react-native-reanimated';
 import { MaterialCommunityIcons as Icon } from '@expo/vector-icons';
 import DateTimePicker from '@react-native-community/datetimepicker';
@@ -21,6 +21,8 @@ import OperationsList from '../components/operations/OperationsList';
 import QuickAddForm from '../components/operations/QuickAddForm';
 import PickerModal from '../components/operations/PickerModal';
 import SearchOverlay from '../components/search/SearchOverlay';
+import SearchBar from '../components/search/SearchBar';
+import FilterChipStrip from '../components/search/FilterChipStrip';
 import * as Currency from '../services/currency';
 import { hasOperation, evaluateExpression } from '../utils/calculatorUtils';
 import useMultiCurrencyTransfer from '../hooks/useMultiCurrencyTransfer';
@@ -31,7 +33,6 @@ import { useSearch } from '../contexts/SearchContext';
 // Note: dynamic createStyles removed to keep linting stable.
 
 const OperationsScreen = () => {
-  console.log('[OperationsScreen] Component rendered');
   const { colors } = useThemeColors();
 
   const { t } = useLocalization();
@@ -41,6 +42,9 @@ const OperationsScreen = () => {
     loading: operationsLoading,
     loadingMore,
     hasMoreOperations,
+    searchState,
+    hasActiveSearch,
+    getSearchFilterCount,
   } = useOperationsData();
   const {
     deleteOperation,
@@ -49,6 +53,8 @@ const OperationsScreen = () => {
     validateOperation,
     loadMoreOperations,
     jumpToDate,
+    setSearchText,
+    updateSearchFilters,
   } = useOperationsActions();
   const { accounts, visibleAccounts } = useAccountsData();
   const { categories } = useCategories();
@@ -65,7 +71,7 @@ const OperationsScreen = () => {
   const [pendingSuggestions, setPendingSuggestions] = useState([]);
   const [filterPanelHeight, setFilterPanelHeight] = useState(0);
 
-  const { searchMode, filtersExpanded } = useSearch();
+  const { searchMode, filtersExpanded, openSearch, closeSearch, reopenSearch, toggleFilters } = useSearch();
   const scrollOffsetRef = useRef(0);
   const prevFiltersExpandedRef = useRef(false);
   const prevSearchModeRef = useRef(searchMode);
@@ -241,15 +247,6 @@ const OperationsScreen = () => {
       }
     }
   }, [pendingScroll, scrollToDateString, operationsLoading, groupedOperations]);
-
-  // Track mount/unmount
-  useEffect(() => {
-    console.log('[OperationsScreen] MOUNTED');
-    return () => {
-      console.log('[OperationsScreen] UNMOUNTING');
-    };
-  }, []);
-
 
   // Auto-populate exchange rate when multi-currency transfer accounts change (async with live rate)
   useEffect(() => {
@@ -636,8 +633,82 @@ const OperationsScreen = () => {
     setModalVisible(false);
   }, []);
 
+  // Search handlers
+  const isSearchOpen = searchMode === 'open';
+
+  const handleCloseSearch = useCallback(() => {
+    closeSearch(hasActiveSearch);
+  }, [closeSearch, hasActiveSearch]);
+
+  const handleToggleFilters = useCallback(() => {
+    toggleFilters();
+  }, [toggleFilters]);
+
+  const handleClearFilterGroup = useCallback((groupKey) => {
+    const clearValues = {
+      text: { text: '' },
+      types: { types: [] },
+      dateRange: { dateRange: { startDate: null, endDate: null } },
+      amountRange: { amountRange: { min: null, max: null } },
+      accountIds: { accountIds: [] },
+      categoryIds: { categoryIds: [] },
+    };
+    updateSearchFilters(clearValues[groupKey]);
+  }, [updateSearchFilters]);
+
+  const handleCollapsedPress = useCallback(() => {
+    if (searchMode === 'collapsed') {
+      const hasOtherFilters =
+        (searchState?.types?.length > 0) ||
+        (searchState?.accountIds?.length > 0) ||
+        (searchState?.categoryIds?.length > 0) ||
+        !!searchState?.dateRange?.startDate ||
+        !!searchState?.dateRange?.endDate ||
+        (searchState?.amountRange?.min !== null && searchState?.amountRange?.min !== undefined) ||
+        (searchState?.amountRange?.max !== null && searchState?.amountRange?.max !== undefined);
+      reopenSearch(searchState?.text !== '', hasOtherFilters, (shouldExpand) => {
+        if (shouldExpand !== filtersExpanded) toggleFilters();
+      });
+    } else {
+      openSearch();
+    }
+  }, [searchMode, searchState, reopenSearch, filtersExpanded, toggleFilters, openSearch]);
+
+  // Back handler for search mode
+  useEffect(() => {
+    if (!isSearchOpen) return;
+    const sub = BackHandler.addEventListener('hardwareBackPress', () => {
+      if (filtersExpanded) {
+        toggleFilters();
+      } else {
+        handleCloseSearch();
+      }
+      return true;
+    });
+    return () => sub.remove();
+  }, [isSearchOpen, filtersExpanded, toggleFilters, handleCloseSearch]);
+
   const quickAddFormComponent = useMemo(() => (
     <>
+      <SearchBar
+        searchText={searchState?.text || ''}
+        onSearchTextChange={setSearchText}
+        onToggleFilters={handleToggleFilters}
+        onClose={handleCloseSearch}
+        filterCount={getSearchFilterCount ? getSearchFilterCount() : 0}
+        colors={colors}
+        t={t}
+        collapsed={!isSearchOpen}
+        onCollapsedPress={handleCollapsedPress}
+      />
+      {isSearchOpen && hasActiveSearch && (
+        <FilterChipStrip
+          searchState={searchState}
+          onClearGroup={handleClearFilterGroup}
+          colors={colors}
+          t={t}
+        />
+      )}
       <Animated.View style={animatedQuickAddClipStyle}>
         <Animated.View style={animatedQuickAddSlideStyle}>
           <QuickAddForm
@@ -670,7 +741,7 @@ const OperationsScreen = () => {
       </Animated.View>
       {filtersExpanded && filterPanelHeight > 0 && <View style={{ height: filterPanelHeight }} />}
     </>
-  ), [animatedQuickAddClipStyle, animatedQuickAddSlideStyle, colors, t, quickAddValues, visibleAccounts, filteredCategories, topCategoriesForType, getCategoryInfo, getAccountName, getAccountBalance, getCategoryName, openPicker, handleQuickAdd, handleAmountChange, handleExchangeRateChange, handleDestinationAmountChange, handleAutoAddWithCategory, topTransferAccountsForForm, handleAutoAddWithAccount, TYPES, rateSource, handleOperationCurrencyChange, foreignRateSource, foreignExchangeRate, filterPanelHeight, filtersExpanded]);
+  ), [animatedQuickAddClipStyle, animatedQuickAddSlideStyle, colors, t, quickAddValues, visibleAccounts, filteredCategories, topCategoriesForType, getCategoryInfo, getAccountName, getAccountBalance, getCategoryName, openPicker, handleQuickAdd, handleAmountChange, handleExchangeRateChange, handleDestinationAmountChange, handleAutoAddWithCategory, topTransferAccountsForForm, handleAutoAddWithAccount, TYPES, rateSource, handleOperationCurrencyChange, foreignRateSource, foreignExchangeRate, filterPanelHeight, filtersExpanded, searchState, setSearchText, handleToggleFilters, handleCloseSearch, getSearchFilterCount, isSearchOpen, handleCollapsedPress, hasActiveSearch, handleClearFilterGroup]);
 
   // Auto-scroll to top when filter panel closes, but only if the user is still
   // near the top (hasn't scrolled into past dates). The threshold is filterPanelHeight:
